requirements.txt
# Multi-App RAG Chatbot Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Environment
python-dotenv==1.0.0

# LangChain & AI
langchain==0.1.0
langchain-community==0.0.13
langchain-text-splitters==0.0.1

# Vector Store
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2

# OpenAI (optional - for LLM)
openai==1.10.0

# Data Validation
pydantic==2.5.3

# HTTP Client (for future integrations)
httpx==0.26.0

# File handling
python-multipart==0.0.6

# Utilities
aiofiles==23.2.1

